Make VB examples match their C# counterparts
Update VB examples to match their C# counterparts.
Add a VB-specific new language feature (multiline string literals) to balance out the C#-specific new language feature (static imports).
Fix a type in the C# `nameof` example - the open parenthesis was to the left of the `if` keyword rather than to the right of it.
Fix the C# static import example to use formatting consistent with what Visual Studio will insert for multiplication.
Change the description of 'using static' to 'static imports' to match existing terminology for the equivalent feature in Java (and possibly other languages). This hopefully makes the feature more searchable and also makes the name more general across both VB and C# ('using static' makes no sense in a VB context).