Add .NET Framework 5B post (#283)
* Add initial post and CSV files

* Test change

* Add CVE, past release and more formatting

There are two CVEs for this release, we have description for 2018-1039 but not for 2018-0765. Also, added format to match 1B release and previous releases.

NOTE: there was a 2C release but we didn't blog about it.  There are new fixes included in 5B that were only first released on 2C.  Should we include these on the Quality and Reliability fixes section?

* Add CVE details

*[] Add bug ID links to each CVE.  Bug query is: https://devdiv.visualstudio.com/DevDiv/_queries?id=1251abbb-99fb-4e0f-bc84-475c1be533a0&_a=query

*[] Add GitHub announcement/issues/XXXX links for each CVE

*[] Determine to add or not add 2C fixes in Quality section.  If so, the query is: https://devdiv.visualstudio.com/DevDiv/_queries?id=25f0b380-b5e8-4de0-b4e8-535e500f919f&_a=query

*[] Add CSV table

*[] Add Docker Image links

* Add updated CSV files

The SO was not in the folder and the MR had updates required

* Add release table for 5B

* Split table

* Update Windows 1803 name

* Remove Windows 10 1511

* Add 'Windows Server'