Raise image threshold to 200kb
This rule is in for some time now and virtually all blog posts end up getting
flagged. It seems 100kb is too restrictive these days and most of our screenshots
are between 100 and 200.

While smaller is obviously better it also seems we'd be better off with rules
that are less noisy.