Remove "Blazor United" naming (#1158)
Previously we used the name "Blazor United" to refer to the various new Blazor improvements in .NET 8 for supporting full stack web UI rendering. But "Blazor United" was not intended to be a new framework or hosting model. Instead it's a collection of new Blazor features that enable full stack web UI development. To avoid future confusion about the name "Blazor United" we've decided to stop using the name and instead focus on the new Blazor features we're adding.