Added PostSlug property and validation rule (#1553)
Added a new nullable string property `PostSlug` to the `BlogFrontMatter` class and a corresponding constant in the `BlogFrontMatterFields` class. Also, a new validation rule has been implemented in the `VR02_MetadataRequired` class to check if the `PostSlug` field is null. If null, an error message is generated using the `MustSpecifyField` method.

Co-authored-by: James Montemagno <james.montemagno@gmail.com>