Update VR03 to allow QuoteBlock as first block (#1694)
The validation rule `VR03_FirstBlockShouldBeParagraph` was updated to allow the first block of a document to be either a `ParagraphBlock` or a `QuoteBlock`. Previously, only `ParagraphBlock` was allowed. Additionally, the warning message was updated to reflect this change, indicating that the first block should be a paragraph with the introduction or an introduction note for guest blog posts.

Co-authored-by: James Montemagno <james.montemagno@gmail.com>